% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_company.R
\name{clean_company}
\alias{clean_company}
\title{Clean company name}
\usage{
clean_company(data)
}
\arguments{
\item{data}{}
}
\value{
dataframe
}
\description{
Because the company name is a free text field and there is often
no one standard way to writing Public Sector names, this function cleans the
errors that have occurred in the data by creating a new column called
`company_cleaned`. Corrections are based on data from HACA 2024 Conference.
}
