% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_eventbrite.R
\name{filter_data}
\alias{filter_data}
\title{Function to filter Eventbrite data}
\usage{
filter_data(
  data,
  tickets,
  remove_dups = TRUE,
  checked_in = c("Both", "Y", "N"),
  country = NA,
  mlcsu = FALSE,
  add_company = FALSE
)
}
\arguments{
\item{data}{extracted from Eventbrite}

\item{tickets}{String Categories are "In-person Ticket  Day 1 - 23rd July
2024", "In-person Ticket Day 2 - 24th July 2024", "Online Ticket" and
"Networking Event".}

\item{remove_dups}{Logical Default is to remove duplicates from emails or
from emails and company_cleaned if used in conjunction with the `add_company`
parameter}

\item{checked_in}{String Default is to include both `Attending` and
`Checked In` or can be defined with `"Both"`. "Y" returns just the tickets
that are `Checked In` and "N" returns just `Attending`.}

\item{country}{String Default is NA and is used in conjunction with
\code{\link{clean_company}} where the country name is added to the cleaned
company: "Northern Ireland", "Scotland", and "Wales".}

\item{mlcsu}{String Default is FALSE. When set to TRUE the company_cleaned
categories for all departments in MLCSU are recoded to "Midlands and
Lancashire CSU". Includes Health Economics Unit, The Strategy Unit and MLCSU.}

\item{add_company}{String Default is FALSE. When TRUE includes the
company_cleaned column in the output data file and is used only where
company_cleaned is needed as there are duplicates as some tickets have the
same email but different organisations.}
}
\value{
Dataframe
}
\description{
Function to filter Eventbrite data
}
